+++ sct/analyses/radiometric_analysis.py
@@ -12,7 +12,12 @@ import logging
 from enum import Enum
 from pathlib import Path
 
-from arepyextras.quality.radiometric_analysis.analysis import gamma_profiles, nesz_profiles, scalloping_profiles
+from arepyextras.quality.core.generic_dataclasses import SARRadiometricQuantity
+from arepyextras.quality.radiometric_analysis.analysis import (
+    average_elevation_profiles,
+    nesz_profiles,
+    scalloping_profiles,
+)
 from arepyextras.quality.radiometric_analysis.config import RadiometricProfilesConfig
 from arepyextras.quality.radiometric_analysis.custom_dataclasses import RadiometricProfilesOutput
 
@@ -27,12 +32,15 @@ class SupportedRadiometricProfiles(Enum):
     """Supported radiometric profiles analyses"""
 
     NESZ = "nesz"
-    GAMMA = "gamma"
+    PROFILES = "average_profiles"
     SCALLOPING = "scalloping"
 
 
 def sct_radiometric_profiles(
-    product_path: str | Path, analysis_type: SupportedRadiometricProfiles, config: RadiometricProfilesConfig
+    product_path: str | Path,
+    analysis_type: SupportedRadiometricProfiles,
+    config: RadiometricProfilesConfig,
+    output_quantity: SARRadiometricQuantity | None = None,
 ) -> list[RadiometricProfilesOutput]:
     """Radiometric profiles SCT wrapper.
 
@@ -44,6 +52,8 @@ def sct_radiometric_profiles(
         type of analysis to perform
     config : RadiometricProfilesConfig
         radiometric profiles configuration
+    output_quantity : SARRadiometricQuantity | None, optional
+        output SAR radiometric quantity, by default None
 
     Returns
     -------
@@ -58,8 +68,8 @@ def sct_radiometric_profiles(
 
     if analysis_type == SupportedRadiometricProfiles.NESZ:
         results = nesz_profiles(product=product, config=config)
-    elif analysis_type == SupportedRadiometricProfiles.GAMMA:
-        results = gamma_profiles(product=product, config=config)
+    elif analysis_type == SupportedRadiometricProfiles.PROFILES:
+        results = average_elevation_profiles(product=product, output_quantity=output_quantity, config=config)
     elif analysis_type == SupportedRadiometricProfiles.SCALLOPING:
         results = scalloping_profiles(product=product, config=config)
 
@@ -94,15 +104,19 @@ def nesz_analysis(
     )
 
 
-def gamma_analysis(
-    product_path: str | Path, config: SCTRadiometricAnalysisConfig | None = None
+def average_elevation_profile_analysis(
+    product_path: str | Path,
+    output_quantity: SARRadiometricQuantity,
+    config: SCTRadiometricAnalysisConfig | None = None,
 ) -> list[RadiometricProfilesOutput]:
-    """SCT Gamma-Zero radiometric block-wise analysis.
+    """SCT Average Radiometric Elevation Profile block-wise analysis.
 
     Parameters
     ----------
     product_path : str | Path
         path to the product to be analyzed
+    output_quantity : SARRadiometricQuantity
+        output SAR radiometric quantity
     config : SCTRadiometricAnalysisConfig | None, optional
         SCT radiometric analysis configuration, by default None
 
@@ -118,7 +132,10 @@ def gamma_analysis(
         config = SCTRadiometricAnalysisConfig()
 
     return sct_radiometric_profiles(
-        product_path=product_path, analysis_type=SupportedRadiometricProfiles.GAMMA, config=config.base_config
+        product_path=product_path,
+        output_quantity=output_quantity,
+        analysis_type=SupportedRadiometricProfiles.PROFILES,
+        config=config.base_config,
     )